DyCE Base Structure Exploration
Playing around with engines referencing engines.

Engines that have a list of properties that each point to an engine.
Engines that select from a list of other engines. And engines that
simply return text. (It has to end somewhere.)

Created a sandbox project for the exploration. Main applications are
taking a back seat for now and may not work properly anymore.